move common types and functions
